+++ src/main.rs
@@ -1,4 +1,5 @@
 #![feature(proc_macro_hygiene, decl_macro)]
+#![feature(uniform_paths)]
 
 #[macro_use]
 extern crate rocket;
@@ -16,35 +17,11 @@ use rocket_contrib::json::Json;
 #[database("my_db")]
 struct MyDatabase(diesel::MysqlConnection);
 
-//mod hero;
-//use hero::Hero;
-
-//use diesel;
-use diesel::prelude::*;
-use diesel::mysql::MysqlConnection;
-
 mod schema;
-use schema::hero as hero_table;
-#[derive(Serialize, Deserialize, Debug, Queryable, Insertable)]
-pub struct Hero {
-    pub id: Option<i32>,
-    pub name: String,
-    pub identity: String,
-    pub hometown: String,
-    pub age: i32,
-}
+mod hero;
+use hero::Hero;
 
-impl Hero {
-    pub fn create(connection: &diesel::MysqlConnection, hero: Json<Hero>) -> Json<Hero> {
-        diesel::insert_into(hero_table::table)
-            .values(&hero)
-            .execute(connection)
-            .expect("Error creating new hero");
-
-        hero_table::table.order(hero_table::id.desc()).first(connection).unwrap()
-    }
-}
-
+#[cfg(test)] mod tests;
 
 #[get("/")]
 fn hello() -> &'static str {
@@ -65,44 +42,4 @@ fn rocket() -> Rocket {
 
 fn main() {
     rocket().launch();
-}
-
-#[cfg(test)]
-mod tests {
-    use super::rocket;
-    use rocket::http::Status;
-    use rocket::local::Client;
-    use crate::Hero;
-    //    use crate::hero::Hero;
-
-    #[test]
-    fn test_hello() {
-        let client = Client::new(rocket()).unwrap();
-        let mut response = client.get("/").dispatch();
-        assert_eq!(response.status(), Status::Ok);
-        assert_eq!(response.body_string(), Some("Hello, world!".into()));
-    }
-
-    #[test]
-    fn it_works() {
-        let hero = Hero {
-            id: Some(1),
-            name: String::from("Superman"),
-            identity: String::from("Clark Kent"),
-            hometown: String::from("Metropolis"),
-            age: 32,
-        };
-        let serialized = serde_json::to_string(&hero).unwrap();
-        println!("serialized = {}", serialized);
-        assert_eq!(serialized, r#"{"id":1,"name":"Superman","identity":"Clark Kent","hometown":"Metropolis","age":32}"#);
-
-        let deserialized: Hero = serde_json::from_str(&serialized).unwrap();
-        println!("deserialized = {:?}", deserialized);
-
-        assert_eq!(deserialized.id, Some(1));
-        assert_eq!(deserialized.name, "Superman");
-        assert_eq!(deserialized.identity, "Clark Kent");
-        assert_eq!(deserialized.hometown, "Metropolis");
-        assert_eq!(deserialized.age, 32);
-    }
 }
